Split up the code into multiple modules
