+++ apps/site/src/routes/events/+page.server.ts
@@ -1,33 +1,5 @@
 import { GOOGLE_CALENDAR_API_KEY } from '$env/static/private';
-
-export interface GoogleCalendarEvent {
-	kind: 'calendar#event';
-	etag: string;
-	id: string;
-	status: 'confirmed' | 'tentative' | 'cancelled';
-	htmlLink: string;
-	created: string;
-	updated: string;
-	summary: string;
-	description?: string;
-	location?: string;
-	creator: { email: string; self?: boolean };
-	organizer: { email: string; self?: boolean };
-	start: {
-		dateTime?: string;
-		date?: string;
-		timeZone?: string;
-	};
-	end: {
-		dateTime?: string;
-		date?: string;
-		timeZone?: string;
-	};
-	iCalUID: string;
-	sequence: number;
-	reminders: { useDefault: boolean };
-	eventType: 'default' | 'outOfOffice' | 'focusTime';
-}
+import { type GoogleCalendarEvent } from '$lib/common/components/Calendar/types';
 
 export const load = async () => {
 	const CALENDAR_IDs = {
